requirements.txt
anyio==3.6.2 ; python_version >= "3.8" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.8" and python_version < "4.0"
charset-normalizer==3.0.1 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
dnspython==2.3.0 ; python_version >= "3.8" and python_version < "4.0"
h11==0.12.0 ; python_version >= "3.8" and python_version < "4.0"
httpcore==0.14.7 ; python_version >= "3.8" and python_version < "4.0"
httpx==0.22.0 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4.0"
jianshuresearchtools==2.10.1 ; python_version >= "3.8" and python_version < "4.0"
jieba==0.42.1 ; python_version >= "3.8" and python_version < "4.0"
jieba[word-split-jieba]==0.42.1 ; python_version >= "3.8" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.8" and python_version < "4.0"
lxml==4.8.0 ; python_version >= "3.8" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.8" and python_version < "4.0"
pillow==9.4.0 ; python_version >= "3.8" and python_version < "4.0"
prettytable==3.6.0 ; python_version >= "3.8" and python_version < "4.0"
pyecharts==2.0.2 ; python_version >= "3.8" and python_version < "4.0"
pymongo==4.3.3 ; python_version >= "3.8" and python_version < "4.0"
pymongo[logging]==4.3.3 ; python_version >= "3.8" and python_version < "4.0"
pypng==0.20220715.0 ; python_version >= "3.8" and python_version < "4.0"
pywebio==1.7.1 ; python_version >= "3.8" and python_version < "4.0"
pywebio[pywebio]==1.7.1 ; python_version >= "3.8" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.8" and python_version < "4.0"
qrcode==7.4.2 ; python_version >= "3.8" and python_version < "4.0"
qrcode[qrcode]==7.4.2 ; python_version >= "3.8" and python_version < "4.0"
rfc3986[idna2008]==1.5.0 ; python_version >= "3.8" and python_version < "4.0"
simplejson==3.18.3 ; python_version >= "3.8" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
sspeedup[logging,pywebio,qrcode,word-split-jieba]==0.2.0 ; python_version >= "3.8" and python_version < "4.0"
tomd==0.1.3 ; python_version >= "3.8" and python_version < "4.0"
tornado==6.2 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.8" and python_version < "4.0"
ua-parser==0.16.1 ; python_version >= "3.8" and python_version < "4.0"
user-agents==2.2.0 ; python_version >= "3.8" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.8" and python_version < "4.0"
